# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * School Library Management System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Book {
  id: number;
  title: string;
  author: string;
  isbn: string;
  genre: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  publisher?: string | null;
  /** @nullable */
  publishedYear?: number | null;
  /** @nullable */
  coverUrl?: string | null;
  totalCopies: number;
  availableCopies: number;
  createdAt: string;
}

export interface BookInput {
  title: string;
  author: string;
  isbn: string;
  genre: string;
  description?: string;
  publisher?: string;
  publishedYear?: number;
  coverUrl?: string;
  totalCopies: number;
}

export interface BookUpdate {
  title?: string;
  author?: string;
  isbn?: string;
  genre?: string;
  description?: string;
  publisher?: string;
  publishedYear?: number;
  coverUrl?: string;
  totalCopies?: number;
}

export interface Student {
  id: number;
  name: string;
  email: string;
  studentId: string;
  /** @nullable */
  phone?: string | null;
  graduationYear: number;
  borrowLimit: number;
  activeLoansCount: number;
  totalFinesOwed?: number;
  createdAt: string;
}

export interface StudentInput {
  name: string;
  email: string;
  studentId: string;
  phone?: string;
  graduationYear: number;
  borrowLimit?: number;
}

export interface StudentUpdate {
  name?: string;
  email?: string;
  phone?: string;
  graduationYear?: number;
  borrowLimit?: number;
}

export type LoanStatus = typeof LoanStatus[keyof typeof LoanStatus];


export const LoanStatus = {
  active: 'active',
  returned: 'returned',
  overdue: 'overdue',
} as const;

export interface Loan {
  id: number;
  studentId: number;
  bookId: number;
  checkedOutAt: string;
  dueDate: string;
  /** @nullable */
  returnedAt?: string | null;
  status: LoanStatus;
  /** @nullable */
  studentName?: string | null;
  /** @nullable */
  bookTitle?: string | null;
  /** @nullable */
  bookAuthor?: string | null;
}

export interface LoanInput {
  studentId: number;
  bookId: number;
  dueDate?: string;
}

export type ReservationStatus = typeof ReservationStatus[keyof typeof ReservationStatus];


export const ReservationStatus = {
  pending: 'pending',
  fulfilled: 'fulfilled',
  cancelled: 'cancelled',
} as const;

export interface Reservation {
  id: number;
  studentId: number;
  bookId: number;
  status: ReservationStatus;
  createdAt: string;
  /** @nullable */
  studentName?: string | null;
  /** @nullable */
  bookTitle?: string | null;
  /** @nullable */
  bookAuthor?: string | null;
}

export interface ReservationInput {
  studentId: number;
  bookId: number;
}

export type ReservationUpdateStatus = typeof ReservationUpdateStatus[keyof typeof ReservationUpdateStatus];


export const ReservationUpdateStatus = {
  pending: 'pending',
  fulfilled: 'fulfilled',
  cancelled: 'cancelled',
} as const;

export interface ReservationUpdate {
  status?: ReservationUpdateStatus;
}

export interface Fine {
  id: number;
  studentId: number;
  loanId: number;
  amount: number;
  reason: string;
  paid: boolean;
  /** @nullable */
  paidAt?: string | null;
  createdAt: string;
  /** @nullable */
  studentName?: string | null;
  /** @nullable */
  bookTitle?: string | null;
}

export interface DashboardSummary {
  totalBooks: number;
  totalStudents: number;
  activeLoans: number;
  overdueLoans: number;
  totalFinesOwed: number;
  booksAvailable: number;
  reservationsPending: number;
}

export interface PopularBook {
  bookId: number;
  title: string;
  author: string;
  loanCount: number;
}

export interface GenreStat {
  genre: string;
  loanCount: number;
}

export type ListBooksParams = {
search?: string;
genre?: string;
available?: boolean;
};

export type ListStudentsParams = {
search?: string;
graduationYear?: number;
};

export type ListLoansParams = {
status?: ListLoansStatus;
studentId?: number;
};

export type ListLoansStatus = typeof ListLoansStatus[keyof typeof ListLoansStatus];


export const ListLoansStatus = {
  active: 'active',
  returned: 'returned',
  overdue: 'overdue',
} as const;

export type ListReservationsParams = {
studentId?: number;
status?: ListReservationsStatus;
};

export type ListReservationsStatus = typeof ListReservationsStatus[keyof typeof ListReservationsStatus];


export const ListReservationsStatus = {
  pending: 'pending',
  fulfilled: 'fulfilled',
  cancelled: 'cancelled',
} as const;

export type ListFinesParams = {
studentId?: number;
paid?: boolean;
};