# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: School Library Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /books:
    get:
      operationId: listBooks
      tags: [books]
      summary: List all books
      parameters:
        - name: search
          in: query
          schema: { type: string }
        - name: genre
          in: query
          schema: { type: string }
        - name: available
          in: query
          schema: { type: boolean }
      responses:
        "200":
          description: List of books
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Book"
    post:
      operationId: createBook
      tags: [books]
      summary: Add a new book
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookInput"
      responses:
        "201":
          description: Created book
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
  /books/{id}:
    get:
      operationId: getBook
      tags: [books]
      summary: Get book by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Book
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
        "404":
          description: Not found
    patch:
      operationId: updateBook
      tags: [books]
      summary: Update a book
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookUpdate"
      responses:
        "200":
          description: Updated book
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
    delete:
      operationId: deleteBook
      tags: [books]
      summary: Delete a book
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ─── Students ─────────────────────────────────────────────────────────────
...
  /loans:
    get:
      operationId: listLoans
      tags: [loans]
      summary: List all active loans
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, returned, overdue]
        - name: studentId
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: List of loans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Loan"
    post:
      operationId: createLoan
      tags: [loans]
      summary: Check out a book to a student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoanInput"
      responses:
        "201":
          description: Created loan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Loan"
  /loans/{id}:
    get:
      operationId: getLoan
      tags: [loans]
      summary: Get loan by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Loan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Loan"
  /loans/{id}/return:
    post:
      operationId: returnLoan
      tags: [loans]
      summary: Return a book
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Updated loan with return info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Loan"

  # ─── Reservations ─────────────────────────────────────────────────────────
...
  /fines:
    get:
      operationId: listFines
      tags: [fines]
      summary: List all fines
      parameters:
        - name: studentId
          in: query
          schema: { type: integer }
        - name: paid
          in: query
          schema: { type: boolean }
      responses:
        "200":
          description: List of fines
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Fine"
  /fines/{id}/clear:
    post:
      operationId: clearFine
      tags: [fines]
      summary: Mark a fine as paid/cleared
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Updated fine
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fine"

  # ─── Dashboard ────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Book:
      type: object
      required: [id, title, author, isbn, genre, totalCopies, availableCopies, createdAt]
      properties:
        id:          { type: integer }
        title:       { type: string }
        author:      { type: string }
        isbn:        { type: string }
        genre:       { type: string }
        description: { type: ["string", "null"] }
        publisher:   { type: ["string", "null"] }
        publishedYear: { type: ["integer", "null"] }
        coverUrl:    { type: ["string", "null"] }
        totalCopies: { type: integer }
        availableCopies: { type: integer }
        createdAt:   { type: string }
    BookInput:
      type: object
      required: [title, author, isbn, genre, totalCopies]
      properties:
        title:       { type: string }
        author:      { type: string }
        isbn:        { type: string }
        genre:       { type: string }
        description: { type: string }
        publisher:   { type: string }
        publishedYear: { type: integer }
        coverUrl:    { type: string }
        totalCopies: { type: integer }
    BookUpdate:
      type: object
      properties:
        title:       { type: string }
        author:      { type: string }
        isbn:        { type: string }
        genre:       { type: string }
        description: { type: string }
        publisher:   { type: string }
        publishedYear: { type: integer }
        coverUrl:    { type: string }
        totalCopies: { type: integer }
...
    Loan:
      type: object
      required: [id, studentId, bookId, checkedOutAt, dueDate, status]
      properties:
        id:           { type: integer }
        studentId:    { type: integer }
        bookId:       { type: integer }
        checkedOutAt: { type: string }
        dueDate:      { type: string }
        returnedAt:   { type: ["string", "null"] }
        status:       { type: string, enum: [active, returned, overdue] }
        studentName:  { type: ["string", "null"] }
        bookTitle:    { type: ["string", "null"] }
        bookAuthor:   { type: ["string", "null"] }
    LoanInput:
      type: object
      required: [studentId, bookId]
      properties:
        studentId: { type: integer }
        bookId:    { type: integer }
        dueDate:   { type: string }
...
    Fine:
      type: object
      required: [id, studentId, loanId, amount, reason, paid, createdAt]
      properties:
        id:          { type: integer }
        studentId:   { type: integer }
        loanId:      { type: integer }
        amount:      { type: number }
        reason:      { type: string }
        paid:        { type: boolean }
        paidAt:      { type: ["string", "null"] }
        createdAt:   { type: string }
        studentName: { type: ["string", "null"] }
        bookTitle:   { type: ["string", "null"] }
...