# lib/api-zod/src/generated/types/bookInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * School Library Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface BookInput {
  title: string;
  author: string;
  isbn: string;
  genre: string;
  description?: string;
  publisher?: string;
  publishedYear?: number;
  coverUrl?: string;
  totalCopies: number;
}