# lib/api-zod/src/generated/types/fine.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * School Library Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface Fine {
  id: number;
  studentId: number;
  loanId: number;
  amount: number;
  reason: string;
  paid: boolean;
  /** @nullable */
  paidAt?: string | null;
  createdAt: string;
  /** @nullable */
  studentName?: string | null;
  /** @nullable */
  bookTitle?: string | null;
}